requirements.txt
Django==5.0.3
django-simple-history==3.5.0
djangorestframework==3.14.0
django-rest-swagger==2.2.0
python-decouple==3.8
psycopg2==2.9.9
shortuuid==1.0.12
drf-yasg==1.21.7

